Modified project files to standardize formatting Two main changes: 1. Added `.prettierrc` file and reformatted all project files 2. Renamed all files that have JSX elements to `*.jsx` instead of `*.js`
