+++ src/components/Meals/Meals.jsx
@@ -2,12 +2,12 @@ import AvailableMeals from "./AvailableMeals";
 import MealsSummary from "./MealsSummary";
 
 const Meals = () => {
-  return (
-    <>
-      <MealsSummary />
-      <AvailableMeals />
-    </>
-  );
+	return (
+		<>
+			<MealsSummary />
+			<AvailableMeals />
+		</>
+	);
 };
 
 export default Meals;
