--- src/components/UI/Card.js
@@ -1,11 +1,0 @@
-import styles from './Card.module.css';
-
-const Card = (props) => {
-    return (
-        <div className={styles.card}>
-         {props.children}   
-        </div>
-    )
-}
-
-export default Card
