+++ src/components/UI/Form/Button.jsx
@@ -1,7 +1,7 @@
 import styles from "./Button.module.css";
 
 const Button = (props) => {
-   return <button className={styles.btn}>{props.children}</button>;
+	return <button className={styles.btn}>{props.children}</button>;
 };
 
 export default Button;
