+++ src/components/custom-hooks/useEventListner.js
@@ -1,21 +1,21 @@
 import { useEffect, useRef } from "react";
 
 export default function useEventListener(
-   eventType,
-   callback,
-   element = window
+	eventType,
+	callback,
+	element = window
 ) {
-   const callbackRef = useRef(callback);
+	const callbackRef = useRef(callback);
 
-   useEffect(() => {
-      callbackRef.current = callback;
-   }, [callback]);
+	useEffect(() => {
+		callbackRef.current = callback;
+	}, [callback]);
 
-   useEffect(() => {
-      if (element == null) return;
-      const handler = (e) => callbackRef.current(e);
-      element.addEventListener(eventType, handler);
+	useEffect(() => {
+		if (element == null) return;
+		const handler = (e) => callbackRef.current(e);
+		element.addEventListener(eventType, handler);
 
-      return () => element.removeEventListener(eventType, handler);
-   }, [eventType, element]);
+		return () => element.removeEventListener(eventType, handler);
+	}, [eventType, element]);
 }
