+++ src/components/custom-hooks/useMediaQuery.js
@@ -2,16 +2,16 @@ import { useState, useEffect } from "react";
 import useEventListener from "./useEventListner";
 
 export default function useMediaQuery(mediaQuery) {
-   const [isMatch, setIsMatch] = useState(false);
-   const [mediaQueryList, setMediaQueryList] = useState(null);
+	const [isMatch, setIsMatch] = useState(false);
+	const [mediaQueryList, setMediaQueryList] = useState(null);
 
-   useEffect(() => {
-      const list = window.matchMedia(mediaQuery);
-      setMediaQueryList(list);
-      setIsMatch(list.matches);
-   }, [mediaQuery]);
+	useEffect(() => {
+		const list = window.matchMedia(mediaQuery);
+		setMediaQueryList(list);
+		setIsMatch(list.matches);
+	}, [mediaQuery]);
 
-   useEventListener("change", (e) => setIsMatch(e.matches), mediaQueryList);
+	useEventListener("change", (e) => setIsMatch(e.matches), mediaQueryList);
 
-   return isMatch;
+	return isMatch;
 }
