+++ src/store/cart-context.js
@@ -1,12 +1,12 @@
 import { createContext } from "react";
 
 const CartContext = createContext({
-   //* default data will not be used but will give auto-completion
-   items: [],
-   totalAmount: 0,
-   addItem: () => {},
-   removeItem: () => {},
-   clearCart: () => {},
+	//* default data will not be used but will give auto-completion
+	items: [],
+	totalAmount: 0,
+	addItem: () => {},
+	removeItem: () => {},
+	clearCart: () => {},
 });
 
 export default CartContext;
